Вливаем ветку с правками в нашу рабочую ветку
