+++ project/libs/post/blog-post/src/index.ts
@@ -3,3 +3,6 @@ export { BlogPostService } from './lib/blog-post.service'
 export { BasePostRepository } from './lib/repositories/base-post.repository'
 export { BasePostFactory } from './lib/factories/base-post.factory'
 export { AllPostRelationRepository } from './lib/repositories/all-post-relation.repository'
+
+export { RepositoryType } from './lib/types/repositories.enum'
+export { FactoryType } from './lib/types/factories.enum'
