+++ project/libs/post/blog-post/src/lib/factories/blog-post-repository.factory.ts
@@ -1,24 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { ModuleRef } from '@nestjs/core';
-import { PostType, PostTypeEnum } from '@project/shared/core'
+import { PostTypeEnum } from '@project/shared/core'
 import { Repository } from '@project/shared/data-access';
 
 import { BasePostEntity } from '../entities/base-post.entity';
-import { PostTextRepository } from '../repositories/post-text.repository';
-import { PostLinkRepository } from '../repositories/post-link.repository';
-import { PostQuoteRepository } from '../repositories/post-quote.repository';
-import { PostPhotoRepository } from '../repositories/post-photo.repository';
-import { PostVideoRepository } from '../repositories/post-video.repository';
-
-const RepositoryType = {
-  [PostType.TEXT]: PostTextRepository,
-  [PostType.LINK]: PostLinkRepository,
-  [PostType.QUOTE]: PostQuoteRepository,
-  [PostType.PHOTO]: PostPhotoRepository,
-  [PostType.VIDEO]: PostVideoRepository,
-} as const;
-
-// type PostRepositoryTypes = (typeof RepositoryType)[keyof typeof RepositoryType];
+import { RepositoryType } from '../types/repositories.enum';
 
 @Injectable()
 export class BlogPostRepositoryFactory {
