+++ project/libs/post/blog-post/src/lib/factories/blog-post.factory.ts
@@ -1,20 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { PostType, PostTypeEnum } from '@project/shared/core';
-import { PostTextFactory } from './post-text.factory';
-import { PostLinkFactory } from './post-link.factory';
-import { PostQuoteFactory } from './post-quote.factory';
-import { PostPhotoFactory } from './post-photo.factory';
-import { PostVideoFactory } from './post-video.factory';
+import { PostTypeEnum } from '@project/shared/core';
+import { FactoryType } from '../types/factories.enum';
 
-const FactoryType = {
-  [PostType.TEXT]: PostTextFactory,
-  [PostType.LINK]: PostLinkFactory,
-  [PostType.QUOTE]: PostQuoteFactory,
-  [PostType.PHOTO]: PostPhotoFactory,
-  [PostType.VIDEO]: PostVideoFactory,
-} as const;
-
-export type PostFactoryTypes = (typeof FactoryType)[keyof typeof FactoryType];
 @Injectable()
 export class BlogPostFactory {
   public getFactoryInstance(postType: PostTypeEnum) {
